Optimize the admin initialization by registering one callback.
Instead of registering multiple callbacks to the `admin_init`, this commit registers `init_admin()`.  Then the new callback method handles the order of calling each of the tasks.

This commit is more performant by O(n-1):

1. Decreases the processing time to do `add_action` as it only have to invoke it once.
2. Decreases the processing time when `do_action` as it only has to invoke one callback.

It's more readable, as we can quickly read that the callback intent is to initialize the admin.