Merge the offline page ID with the current "post__not_in" var.
Being mindful of performance, this commit checks if the "post__not_in" has a value.  If yes, then it merges with the offline page ID and then filters for only the unique values.

Else, it sets the offline ID only, thereby skipping the `array_merge` and `array_unique`.

The code is more complex.  But it serves to only run the the array functions when needed.