Eliminate custom WPRouter in part to ensure precache route is added first
This change unfortunately means that there will not be a Workbox notice raised when there are more than one registered route that matches a given URL. But the benefits are that it is easier to fix the precaching conflict problem and it eliminates needing to include a copy of the registerNavigationRoute function.

It would be beneficial to propose a change to Workbox to report multiple route matches when the Workbox logging level is set appropriately.