Output link/meta tags depending on web app manifest display value
This changes when the (apple-)mobile-web-app-capable meta tags (and
apple-touch-startup-image link tag) are printed to match the display
mode set in the web app manifest.

This also removes the apple-touch-fullscreen meta tag, as this tag has
the same effect as the apple-mobile-web-app-capable meta tag but is not
included in any Safari documentation.